relocation to primitive 1
relocation to primitive 2

relocation to primitive 3

relocation to primitive 4

relocation to primitive 5
