+++ lib/wallet/src/Cardano/Wallet.hs
@@ -551,6 +551,10 @@ import Cardano.Wallet.Primitive.Types.DRep
 import Cardano.Wallet.Primitive.Types.Hash
     ( Hash (..)
     )
+import Cardano.Wallet.Primitive.Types.MetadataEncryption
+    ( ErrMetadataDecryption
+    , ErrMetadataEncryption
+    )
 import Cardano.Wallet.Primitive.Types.Range
     ( Range (..)
     )
@@ -859,7 +863,6 @@ import qualified Cardano.Wallet.Primitive.Types.Tx.TxOut as TxOut
 import qualified Cardano.Wallet.Primitive.Types.UTxO as UTxO
 import qualified Cardano.Wallet.Primitive.Types.UTxOStatistics as UTxOStatistics
 import qualified Cardano.Wallet.Read as Read
-import qualified Cryptography.Cipher.AES256CBC as AES256CBC
 import qualified Data.ByteArray as BA
 import qualified Data.Delta.Update as Delta
 import qualified Data.Foldable as F
@@ -3755,8 +3758,7 @@ data ErrConstructTx
     | ErrConstructTxDelegationInvalid
     | ErrConstructTxVotingInWrongEra
     | ErrConstructTxWithdrawalWithoutVoting
-    | ErrConstructTxIncorrectRawMetadata
-    | ErrConstructTxEncryptMetadata AES256CBC.CipherError
+    | ErrConstructTxFromMetadataEncryption ErrMetadataEncryption
     | ErrConstructTxNotImplemented
     deriving (Show, Eq)
 
@@ -3767,14 +3769,8 @@ data ErrGetPolicyId
     deriving (Show, Eq)
 
 -- | Errors that can occur when decoding a transaction.
-data ErrDecodeTx
-    = ErrDecodeTxMissingMetadataKey
-    | ErrDecodeTxMissingEncryptionMethod
-    | ErrDecodeTxMissingValidEncryptionPayload
-    | ErrDecodeTxDecryptedPayload Text
-    | ErrDecodeTxMissingSalt
-    | ErrDecodeTxDecryptPayload AES256CBC.CipherError
-    | ErrDecodeTxEncryptedPayloadWrongBase
+newtype ErrDecodeTx
+    = ErrDecodeTxFromMetadataDecryption ErrMetadataDecryption
     deriving (Show, Eq)
 
 -- | Errors that can occur when signing a transaction.
